fix bigint to int

+++ python-api/features/tables/category/models.py
@@ -22,7 +22,7 @@ class CategoryModel(Base):
     __default_sort__ = "id"
 
     id: Mapped[int] = mapped_column(
-        sqlalchemy.BigInteger,
+        sqlalchemy.Integer,
         primary_key=True, 
         autoincrement=True,
         unique=True,
